Refactoring done, implementing task list

+++ source/math/arpack_extra/matrix_product_stl.cpp
@@ -57,8 +57,8 @@ StlMatrixProduct<Scalar>::StlMatrixProduct(
     A_ptr(A_) ,L(L_), form(form_), side(side_)
 {
     if (copy_data){
-        A_stl.resize(L*L);
-        std::copy(A_ptr,A_ptr + L*L, A_stl.begin());
+        A_stl.resize(static_cast<size_t>(L*L));
+        std::copy(A_ptr,A_ptr + static_cast<size_t>(L*L), A_stl.begin());
         A_ptr = A_stl.data();
     }
     stl_lu::init<Scalar>();
